Updated Readme to be Admin ready, added default layout
While I want to start the project in as close to an "Empty project"
state as reasonable, I don’t want the project to touch CSS or JS any
more than needed as that isn’t the focus of the project. Despite HTML
not being a core focus of the project, it’s necessary the learner will
need to do a decent amount of HTML to establish an understanding of
razor, tag-helpers, etc.